pyproject.toml
[tool.poetry]
name = "edu-meet-admin"
version = "0.1.0"
description = ""
authors = ["Asgef <asgefes1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.3"
gunicorn = "^23.0.0"
python-dotenv = "^1.0.1"
dj-database-url = "^2.3.0"
psycopg2-binary = "^2.9.10"
whitenoise = "^6.8.2"
django-bootstrap5 = "^24.3"
django-scheduler = {git = "https://github.com/Asgef/django-scheduler.git", rev = "chore/565-update-models"}

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
